+++ python/cudf/tests/test_groupby.py
@@ -355,3 +355,24 @@ def test_groupby_column_numeral():
     pxx = p[0].sum()
     gxx = g[0].sum()
     assert_eq(pxx, gxx)
+
+
+@pytest.mark.parametrize('series', [[0, 1, 0], [1, 1, 1], [0, 1, 1], [1, 2, 3], [4, 3, 2], [0, 2, 0]])  # noqa: E501
+def test_groupby_external_series(series):
+    pdf = pd.DataFrame({'x': [1., 2., 3.], 'y': [1, 2, 1]})
+    gdf = DataFrame.from_pandas(pdf)
+    pxx = pdf.groupby(pd.Series(series)).x.sum()
+    gxx = gdf.groupby(cudf.Series(series)).x.sum()
+    assert_eq(pxx, gxx)
+
+
+@pytest.mark.xfail(raises=NotImplementedError,
+                   reason="CUDF doesn't support arbitrary series index lengths"
+                          "for groupby")
+@pytest.mark.parametrize('series', [[0, 1], [1, 1, 1, 1]])
+def test_groupby_external_series_incorrect_length(series):
+    pdf = pd.DataFrame({'x': [1., 2., 3.], 'y': [1, 2, 1]})
+    gdf = DataFrame.from_pandas(pdf)
+    pxx = pdf.groupby(pd.Series(series)).x.sum()
+    gxx = gdf.groupby(cudf.Series(series)).x.sum()
+    assert_eq(pxx, gxx)