+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - PR #906 Add binary and comparison ops to DataFrame
 - PR #958 Support unary and binary ops on indexes
 - PR #964 Add `rename` method to `DataFrame`, `Series`, and `Index`
+- PR #990 Add external series groupby support
 - PR #988 Add top-level merge function to cuDF
 - PR #992 Add comparison binaryops to DateTime columns
 - PR #996 Replace relative path imports with absolute paths in tests
@@ -68,6 +69,7 @@
 - PR #959 Groupby: Problem with column name lookup
 - PR #950 Converting dataframe/recarry with non-contiguous arrays
 - PR #963 CSV Reader: Fix another issue with missing data rows when using byte_range
+
 
 # cuDF 0.5.1 (05 Feb 2019)